refactor(accessibility): extract pkg/accessibility for library + CLI reuse
Move all accessibility detection and score-modifier logic from the
cmd/titus package (package main) into a new importable package
pkg/accessibility. The CLI wrapper in cmd/titus/accessibility.go
becomes a thin re-export shim so existing CLI tests and scan.go
continue to compile without changes.
